Configuration updates
Configures a GeoServer password via `.env`.  To make this happen, the `DJANGO_SETTINGS_MODULE`
must not be configured at the service template as it gets _overridden_
not merged ^^. In addition to that, the `OGC_SERVER_DEFAULT_*`
parameters should not be overridden by the settings.py, so we reset
them after loading the `geonode.settings` module.